+++ src/main/kotlin/aoc/template.kt
@@ -5,16 +5,16 @@ import kotlin.math.*
 
 fun main() {
     check(execute1(readTestLines(1)), -1)
-    execute1(readLines()).let { println(it) } // ; check(it, readAnswerAsInt(1)) }
+    execute1(readLines()).let { println(it) } // ; check(it, readAnswerAsLong(1)) }
 
     // check(execute2(readTestLines(1)), -1)
-    // execute2(readLines()).let { println(it) } // ; check(it, readAnswerAsInt(2)) }
+    // execute2(readLines()).let { println(it) } // ; check(it, readAnswerAsLong(2)) }
 }
 
-private fun execute1(input: List<String>): Int {
-    return 0
+private fun execute1(input: List<String>): Long {
+    return 0L
 }
 
-private fun execute2(input: List<String>): Int {
-    return 0
+private fun execute2(input: List<String>): Long {
+    return 0L
 }
